Factor out from_xml from bpmn
