+++ src/bpmncwpverify/core/bpmn.py
@@ -4,8 +4,6 @@ from xml.etree.ElementTree import Element
 from bpmncwpverify.core.state import State
 from returns.result import Result
 from abc import abstractmethod
-from returns.pipeline import is_successful
-from returns.functions import not_
 from bpmncwpverify.core.error import (
     BpmnStructureError,
     Error,
@@ -372,44 +370,6 @@ class Bpmn:
         self.accept(promela_visitor)
 
         return str(promela_visitor)
-
-    @staticmethod
-    def from_xml(root: Element, symbol_table: State) -> Result["Bpmn", Error]:
-        from bpmncwpverify.builder.bpmn_builder import BpmnBuilder
-
-        ##############
-        # Build and add processes
-        ##############
-        processes = root.findall("bpmn:process", BPMN_XML_NAMESPACE)
-        bpmn_builder = BpmnBuilder()
-        for process_element in processes:
-            process = Process.from_xml(process_element, symbol_table)
-            if not_(is_successful)(process):
-                return cast(Result[Bpmn, Error], process)
-            bpmn_builder = bpmn_builder.with_process(process.unwrap())
-
-        ##############
-        # Build and add messages
-        ##############
-        collab = root.find("bpmn:collaboration", BPMN_XML_NAMESPACE)
-        if collab is not None:
-            # TODO: write test for messages in the bpmn diagram
-            bpmn_builder = bpmn_builder.with_process_elements()
-            for msg_flow in collab.findall("bpmn:messageFlow", BPMN_XML_NAMESPACE):
-                source_ref, target_ref = (
-                    msg_flow.get("sourceRef"),
-                    msg_flow.get("targetRef"),
-                )
-
-                message = MessageFlow.from_xml(msg_flow)
-
-                assert source_ref and target_ref
-
-                bpmn_builder = bpmn_builder.with_message(
-                    message, source_ref, target_ref
-                )
-
-        return cast(Result[Bpmn, Error], bpmn_builder.build())
 
 
 ###################
